fix: move header component to ui folder
