+++ src/widgets/Header/index.ts
@@ -1,1 +1,1 @@
-export { Header } from "./Header";
+export { Header } from "./ui/Header";
