+++ src/widgets/Header/ui/HeaderDrawer/HeaderDrawer.tsx
@@ -9,17 +9,17 @@ import {
   ListItem,
   ListItemText,
 } from "@mui/material";
-import { navigationItems } from "../model/items";
+import { navigationItems } from "../../model/items";
 import LoginIcon from "@mui/icons-material/Login";
 import CloseIcon from "@mui/icons-material/Close";
 import ShoppingCartOutlinedIcon from "@mui/icons-material/ShoppingCartOutlined";
 import AccountCircleIcon from "@mui/icons-material/AccountCircle";
 import { listItemTextStyle, rootStyle } from "./style";
-import { Link } from "../../../shared/ui/Link";
-import { Logo } from "../../../shared/ui/Logo";
-import { ThemeColors } from "../../../shared/constants/colors";
+import { Link } from "../../../../shared/ui/Link";
+import { Logo } from "../../../../shared/ui/Logo";
+import { ThemeColors } from "../../../../shared/constants/colors";
 import { useNavigate } from "react-router-dom";
-import { Paths } from "../../../shared/constants/paths";
+import { Paths } from "../../../../shared/constants/paths";
 
 interface HeaderDrawerProps {
   isOpen: boolean;
